second impact

+++ src/main/kotlin/org/snd/image/Dimension.kt
@@ -1,6 +1,6 @@
 package org.snd.image
 
 data class Dimension(
-    val width: Int,
-    val height: Int,
+    val width: Float,
+    val height: Float,
 )
